requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.4

# OpenAI
openai>=1.6.0

# PDF Processing
pdfplumber>=0.10.0

# Analytics / Statistics
scipy>=1.11.0
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Telephony
twilio>=8.0.0

# Utilities
httpx>=0.25.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
